+++ spark-bigquery-connector-common/src/main/java/com/google/cloud/spark/bigquery/SparkBigQueryConfig.java
@@ -294,7 +294,7 @@ public class SparkBigQueryConfig
         spark.sparkContext().hadoopConfiguration(),
         customDefaults,
         spark.sparkContext().defaultParallelism(),
-        spark.sqlContext().conf(),
+        spark.sessionState().conf(),
         spark.version(),
         schema,
         tableIsMandatory,