+++ spark-bigquery-dsv2/spark-4.0-bigquery/src/test/java/com/google/cloud/spark/bigquery/integration/Spark40OpenLineageIntegrationTest.java
@@ -15,4 +15,7 @@
  */
 package com.google.cloud.spark.bigquery.integration;
 
+import org.junit.Ignore;
+
+@Ignore("OL DatabricksEventFilter throws exception in spark 4")
 public class Spark40OpenLineageIntegrationTest extends OpenLineageIntegrationTestBase {}