+++ CHANGES.md
@@ -6,6 +6,7 @@
 * PR #1367: Query Pushdown is no longer supported.
 * PR #1369: Catalog enhancements
 * PR #1370: Add backwards compatibility for OpenLineage prior to version 1.18.0.
+* PR #1374: Ensure TableId includes project ID if not explicitly set
 
 ## 0.42.2 - 2025-05-16
 * PR #1347: Get lineage out of query. Thanks @ddebowczyk92