+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## Next
+## 0.42.4 - 2025-06-20
 * PR #1380: Fixed ImpersonatedCredentials serialization
 * PR #1381: Added the option to set custom credentials scopes
 * PR #1384: Support reading from BigQuery tables with spaces in their names